Don't use Permutation<4>::transposition
For permutations of 4 items, the fully explicit constructor is actually
less verbose.

+++ src/permutation.hpp
@@ -40,7 +40,7 @@ Permutation<N>::Permutation() {
 template <size_t N>
 template <typename... T>
 Permutation<N>::Permutation(T... args)
-    : v_{args...}
+    : v_{size_t(args)...}
 {
 }
